requirements.txt
psycopg2==2.9.9
fastapi==0.115.2
scikit-learn==1.5.2
tensorflow==2.17.0
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
